src/booking_agent/requirements.txt
bedrock-agentcore
nova-act
pyyaml
boto3
structlog
python-dotenv
pydantic
python-dateutil
